Move generic error types to pagopa-interop-models
