+++ packages/catalog-process/src/model/types.ts
@@ -1,8 +1,14 @@
 import { ZodiosBodyByPath, ZodiosErrorByPath } from "@zodios/core";
 import { P, match } from "ts-pattern";
-import { CatalogProcessError, ErrorTypes } from "pagopa-interop-models";
+import {
+  CatalogProcessError,
+  ErrorTypes,
+  Problem,
+  makeApiProblem,
+} from "pagopa-interop-models";
 import { api } from "./generated/api.js";
 
+const servicePrefix = "catalog";
 type Api = typeof api.api;
 export type ApiEServiceSeed = ZodiosBodyByPath<Api, "post", "/eservices">;
 
@@ -40,42 +46,6 @@ export type ApiError =
   | ApiErrorNameConflict
   | ApiInternalServerError
   | Problem;
-
-export type ProblemError = {
-  code: string;
-  detail: string;
-};
-
-export type Problem = {
-  type: string;
-  status: number;
-  title: string;
-  correlationId?: string;
-  detail: string;
-  errors: ProblemError[];
-};
-
-export function makeApiProblem(
-  errorCode: string,
-  httpStatus: number,
-  title: string,
-  detail: string
-): Problem {
-  return {
-    type: "https://docs.pagopa.it/interoperabilita-1/", // TODO change this with properly schema definition URI
-    title,
-    status: httpStatus,
-    detail,
-    errors: [
-      {
-        code: errorCode,
-        detail,
-      },
-    ],
-  };
-}
-
-const servicePrefix = "catalog";
 
 export function makeApiError(error: unknown): ApiError {
   return match<unknown, ApiError>(error)
